+++ src/app/components/ManageThemesModal/ThemeGroupDropDownMenu.tsx
@@ -8,8 +8,8 @@ import {
   DropdownMenuTrigger,
   DropdownMenuContent,
   DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuItemIndicator,
+  ScrollDropdownMenuRadioItem,
+  ScrollDropdownMenuItemIndicator,
   DropdownMenuSeparator,
 } from '../DropdownMenu';
 import { styled } from '@/stitches.config';
@@ -22,18 +22,6 @@ type Props = {
   onChange: (value: string) => void
   addGroup: () => void
 };
-
-const StyledDropdownMenuItemIndicator = styled(DropdownMenuItemIndicator, {
-  position: 'relative',
-  left: 0,
-});
-
-const StyledDropdownMenuRadioItem = styled(DropdownMenuRadioItem, {
-  padding: '$2 $3 $2 $2',
-  userselect: 'none',
-  display: 'flex',
-  alignItems: 'center',
-});
 
 const StyledDropdownMenuTrigger = styled(DropdownMenuTrigger, {
   padding: 0,
@@ -49,21 +37,21 @@ export const ThemeGroupDropDownMenu: React.FC<Props> = ({
   const themeGroupList = React.useMemo(() => availableGroups.map((groupName) => {
     const handleSelect = () => handleSelectGroup(groupName);
     return (
-      <StyledDropdownMenuRadioItem
+      <ScrollDropdownMenuRadioItem
         key={groupName}
         value={groupName}
           // eslint-disable-next-line react/jsx-no-bind
         onSelect={handleSelect}
       >
         <Box css={{ width: '$5', marginRight: '$2' }}>
-          <StyledDropdownMenuItemIndicator>
+          <ScrollDropdownMenuItemIndicator>
             <CheckIcon />
-          </StyledDropdownMenuItemIndicator>
+          </ScrollDropdownMenuItemIndicator>
         </Box>
         <Box>
           {groupName}
         </Box>
-      </StyledDropdownMenuRadioItem>
+      </ScrollDropdownMenuRadioItem>
     );
   }), [availableGroups, handleSelectGroup]);
